+++ Model/ConfigProvider.php
@@ -245,7 +245,7 @@ class ConfigProvider extends CcGenericConfigProvider
                     'userTokenId'           => $this->moduleConfig->getQuoteBillingAddress()['email'],
                     'applePayLabel'         => $this->moduleConfig->getConfigValue('apple_pay_label', 'web_sdk'),
                     'currencyCode'          => $this->moduleConfig->getQuoteBaseCurrency(), 
-					'style'					=> json_decode($this->moduleConfig->getConfigValue('sdk_style', 'basic'), true),
+					'style'					=> json_decode((string) $this->moduleConfig->getConfigValue('sdk_style', 'basic'), true),
                 ],
             ],
         ];